[paimon] enforce query dialect representation for view conversion
